app/requirements.txt
# Core
fastapi==0.124.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20

# Database
motor==3.7.1
pymongo==4.15.5

# Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
email-validator==2.3.0

# Pydantic
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.12.0

# Storage
boto3==1.34.0
minio==7.2.1

# Image Processing
Pillow==10.4.0

# Utilities
python-dateutil==2.8.2
pytz==2025.2
python-magic==0.4.27

requests==2.32.5

# Development
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.1
